+++ src/client/src/modules/inspector/components/tabs/InspectorTabsComponent.tsx
@@ -6,12 +6,12 @@ import { ShouldShowTabs } from "./helpers";
 import { InspectorElement } from "../../types";
 import { GetAttributesElement } from "./components/parameters/helpers/GetAttributesElement";
 import {
-  InspectorTabWrapper,
+  InspectorTabElement,
   AdminComponent,
   AttributesComponent,
   RelationsComponent,
   SimpleTypesComponent,
-  TerminalsAttributesComponent,
+  TerminalAttributesComponent,
 } from "./components";
 
 interface Props {
@@ -27,7 +27,12 @@ interface Props {
   isOffPage: boolean;
 }
 
-export const InspectorTabs = ({
+/**
+ * Component for all the tabs showed in the Inspector Module.
+ * @param props
+ * @returns all available tabs.
+ */
+export const InspectorTabsComponent = ({
   project,
   element,
   activeTabIndex,
@@ -40,16 +45,12 @@ export const InspectorTabs = ({
   isOffPage,
 }: Props) => {
   const shouldShowTabs = ShouldShowTabs(element);
+  const elements = GetAttributesElement(element);
 
   const tabs = [
     <AdminComponent key={0} element={element} project={project} />,
-    <AttributesComponent
-      key={1}
-      attributesElem={GetAttributesElement(element)}
-      inspectorParentElem={element}
-      attributeItems={attributes}
-    />,
-    <TerminalsAttributesComponent key={2} element={element} terminals={terminals} />,
+    <AttributesComponent key={1} attributesElem={elements} inspectorParentElem={element} attributeItems={attributes} />,
+    <TerminalAttributesComponent key={2} element={element} terminals={terminals} />,
     <RelationsComponent key={3} element={element} />,
     <SimpleTypesComponent key={4} element={element} simpleItems={simpleItems} />,
   ];
@@ -60,7 +61,7 @@ export const InspectorTabs = ({
         tabs.map(
           (tab, i) =>
             shouldShowTabs[i] && (
-              <InspectorTabWrapper
+              <InspectorTabElement
                 key={i}
                 element={element}
                 index={i}
@@ -72,7 +73,7 @@ export const InspectorTabs = ({
                 nodes={project?.nodes}
               >
                 {tab}
-              </InspectorTabWrapper>
+              </InspectorTabElement>
             )
         )}
     </>
