+++ src/client/src/modules/inspector/components/tabs/components/InspectorTabElement.styled.ts
